+++ pyof/v0x01/controller2switch/port_mod.py
@@ -37,9 +37,12 @@ class PortMod(GenericMessage):
             hw_addr (HWAddress): The hardware address is not configurable.
                 This is used to sanity-check the request,
                 so it must be the same as returned in an ofp_phy_port struct.
-            config (PortConfig): Bitmap of OFPPC_* flags
-            mask (PortConfig): Bitmap of OFPPC_* flags to be changed
-            advertise (PortFeatures): Bitmap of "ofp_port_features"s
+            config (~pyof.v0x01.common.phy_port.PortConfig):
+                Bitmap of OFPPC_* flags
+            mask (~pyof.v0x01.common.phy_port.PortConfig):
+                Bitmap of OFPPC_* flags to be changed
+            advertise (~pyof.v0x01.common.phy_port.PortFeatures):
+                Bitmap of "ofp_port_features"s
         """
         super().__init__(xid)
         self.port_no = port_no